Flurry of changes all over the place
Nearly working end to end. Clearer separation of concerns
between finding a template and filling a template

Evaluating the data expression passed to datadoc in a clean
environment to make sure it evaluates in an Rmd file

Need to decide behaviour of `dir` arg for generating bookdown book in the
target directory.

get_template is referencing templates that don't exist yet...

Better error messages all around